Fix menu toolbar actions
Now when the sub windows (except for the ones that are used to create a
new game) in the main window are closed by the user, they are just
hidden instead of being destroyed, so that all the data previously
inserted by the user are conserved.

By clicking one of the actions in the Window menu, a sub window that was
previously hidden is shown again; if it is already open, nothing happen.